SCB-155 Change the package of metric.go
